"Updated color variables and applied them throughout the stylesheets; added consistent border radius and box-shadow to elements; updated image styling"
This commit message describes the changes made in the code diff, including the update of color variables and their application throughout the stylesheets, as well as the addition of consistent border radius and box-shadow to elements and updated image styling. It is concise and accurately describes the changes made in the code.