'Update cursor style for buttons and add hover effect for links in FAQ and index pages'
This commit includes changes to the FAQ.css and index.css files. In FAQ.css, the cursor style for the button in the fourth section has been changed from 'help' to 'pointer'. Additionally, a new rule has been added to style the hover effect for links in the last section of index.css. The hover effect changes the color of the link text. Finally, another rule has been added to maintain the color of the icon when the link is hovered over.